Loading and parsing of Remote versions. Loading and parsing of items and annotations. Batch Sync Engine.
Refactoring of Network parsing.
CustomResult for Success & Network & Code errors
Custom JSON Parsing for:
    CreatorResponse
    ItemResponse
    LibraryResponse
    LinkResponse
    LinksResponse
    TagResponse
    UserResponse

Versions & Objects Network requests
InitializeCustomLibrariesDbRequest
SyncVersionsDbRequest
StoreItemsResponse
SyncVersionsSyncAction

Error handling

+++ app/src/main/java/org/zotero/android/sync/LibraryIdentifier.kt
@@ -14,7 +14,7 @@ sealed class LibraryIdentifier {
             }
         }
 
-    fun apiPath(userId: Int): String {
+    fun apiPath(userId: Long): String {
         return when (val q = this) {
             is group ->
                 "groups/${q.groupId}"
